# go-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "cows.shmyaks.ru API",
        "title": "AgroCode Service API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/v1/login": {
            "post": {
                "description": "Авторизует пользователя с предоставленным email и паролем.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Логин",
                "parameters": [
                    {
                        "description": "Запрос на авторизацию пользователя",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Пользователь авторизован",
                        "schema": {
                            "$ref": "#/definitions/model.CodeResponse"
                        }
                    },
                    "400": {
                        "description": "Не удалось авторизовать пользователя",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/user/me": {
            "get": {
                "description": "Возвращает все данные о пользователе по идентификатору или email из cookie.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Получить данные пользователя",
                "responses": {
                    "200": {
                        "description": "Данные пользователя успешно получены",
                        "schema": {
                            "$ref": "#/definitions/model.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка при получении данных пользователя",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/user/{id}": {
            "get": {
                "description": "Возвращает все данные о пользователе по идентификатору или email из cookie.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Получить данные пользователя",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID пользователя, если доступ по ID",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные пользователя успешно получены",
                        "schema": {
                            "$ref": "#/definitions/model.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка при получении данных пользователя",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/user_create": {
            "post": {
                "description": "Создает нового пользователя с предоставленным email, паролем и именем.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Создать нового пользователя",
                "parameters": [
                    {
                        "description": "Запрос на создание пользователя",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Пользователь успешно создан",
                        "schema": {
                            "$ref": "#/definitions/model.CodeResponse"
                        }
                    },
                    "400": {
                        "description": "Не удалось создать пользователя",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.CodeResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "model.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "model.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "model.UserCreateRequest": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "model.UserResponse": {
            "type": "object",
            "properties": {
                "animals": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "photo": {
                    "type": "string"
                },
                "rating": {
                    "type": "number"
                },
                "reviews": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "reviews_count": {
                    "type": "integer"
                },
                "stats_matches": {
                    "type": "integer"
                },
                "stats_views": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "CookieAuth": {
            "type": "apiKey",
            "name": "Authtoken",
            "in": "cookie"
        }
    }
}